requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
certifi==2018.1.18
chardet==3.0.4
Django==2.0.3
djangorestframework==3.7.7
idna==2.6
kombu==3.0.37
psycopg2==2.7.4
pymongo==3.6.1
pytz==2018.3
requests==2.18.4
urllib3==1.22
vine==1.1.4
